+++ BearishEngine/src/Core/Actor.cpp
@@ -11,11 +11,11 @@ Actor::Actor(Transform transform) {
 }
 
 Actor::~Actor() {
-	for (i32 i = 0; i < _children.size(); i++) {
+	for (i32 i = 0; i < (i32)_children.size(); i++) {
 		delete _children.at(i);
 	}
 
-	for (i32 i = 0; i < _components.size(); i++) {
+	for (i32 i = 0; i < (i32)_components.size(); i++) {
 		delete _components.at(i);
 	}
 }
